split cache policies test suites into different files
